requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black, fiona
black==19.10b0            # via -r requirements.in
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click-plugins==1.1.1      # via fiona
click==7.0                # via black, click-plugins, cligj, fiona, pip-tools
cligj==0.5.0              # via fiona
fiona==1.8.13.post1       # via geopandas
geopandas==0.7.0          # via -r requirements.in
idna==2.9                 # via requests
munch==2.5.0              # via fiona
mwparserfromhell==0.5.4   # via -r requirements.in
numpy==1.18.3             # via pandas
pandas==1.0.3             # via geopandas
pathspec==0.8.0           # via black
pillow==7.0.0             # via -r requirements.in
pip-tools==4.5.1          # via -r requirements.in
psycopg2==2.8.4           # via -r requirements.in
pyahocorasick==1.4.0      # via -r requirements.in
pycountry==19.8.18        # via -r requirements.in
pyproj==2.6.0             # via geopandas
python-dateutil==2.8.1    # via pandas
pytz==2019.3              # via pandas
pyyaml==5.3.1             # via -r requirements.in
regex==2020.4.4           # via black
requests==2.23.0          # via -r requirements.in
rtree==0.9.4              # via -r requirements.in
shapely==1.7.0            # via -r requirements.in, geopandas
six==1.14.0               # via fiona, munch, pip-tools, python-dateutil
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
urllib3==1.25.8           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
